Adding support for batch search Added the batch search feature. Created batch activesupport object. Created a basic view file for the batch (this is going to need some work latter).
Added support for tiff file attachments in web_steps.rb

Added a scenario for a batch search for a batch that can't be found.
Added a batch_step to test that the batch can't be found.
Put a begin rescue block in the show method to render a 404 page.

Moved the sequencescape_uri constant to a Settings singleton.
Set up a Setting singleton initializer.
Moved settings to config/setting/ directory with a YAML file or symlink for each environment.