[fixes 3484028 3487351] Fixes issues with bulk uploads
Although the YUI Uploader supports sending multiple files it seems to
get very confused when you request to send them all, and will sometimes
fail them with a cryptic error message.  The only way to solve this is
to send the files sequentially.

Once a bulk upload has failed it is important that any resumption
overwrites what was previously there, otherwise the images get
completely screwed.