Ensure that find_in_batches errors for mismatched size.
When a batch of images has less entries than the requested limit, do a
second check just to see if there is a problem.  If the size differs
between the two batches then we raise an error because this is bad.

Also, we're ordering by created_at *and* id to ensure that order clashes
don't produce any odd behaviour.