Monkey patch for MySQL support
The entire application is full of deprecation errors and ancient
gems. We'll need to give it a full update before long as its not
as legacy as we'd like. These patches are to resolve the immediate
issues posed by the MySQL 5.7 upgrade, but Clusterview should
be the first app that gets propely updated.

- Ruby Gems 2 monkey patch for Rails 2
- MySQL 5.7 monkey patch