Batch controller tests use actual file
Previously the tests were failing due to an ActiveRecord::RecodInvalid
exception raised by Paperclip when identify failed to recognise the file
type. This seems unrelated to the 5.7 changes and its unclear exactly
when it became broken. Given we've already had to add monkey patches
for two external dependencies though, a git bisect would be painful,
and possibly wouldn't even find the issue if it was a result of changes
to imagemagik or identify.